Fixes client up/down, left/right, front/back, clockwise/counterclockwise functions to accept 0 speed. Adds test. Resolves #106
Previous implementation would not allow one to supply 0 for speed to stop moving along that axis. The only way to stop would be to issue a client.stop(). But that would also cancel any other movement going on at the time.
For example, centering a joystick should stop pitch, but should not stop yaw being controlled by a dpad on the same controller.